text_message/requirements.txt
certifi==2022.12.7
chardet==4.0.0
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
pyparsing==2.4.7
python-dateutil
python-dotenv
requests
roboflow>=1.0.5
six
urllib3==1.26.6
wget
tqdm>=4.41.0
PyYAML>=5.3.1
requests_toolbelt
pandas
opencv-python-headless
twilio